Reorganize into Outline Style Guide
This reorganizes the style guide entries into an outline, grouped by "Clarity" (defined as makes no impact on code logic) and "Usage" (will affect program). The outline is intended to have all information necessary for quick reference. The outline is the style guide, not a table of contents. The (still crucial) 
 
An undifferentiated list is hard to process, retain information from, and refer to. Furthermore, some patterns in use of verbs appeared, such as "prefer, omit, avoid" that seem to fit with different sections; a language usage definitions for the style guide may be warranted.